:sparkles: feat: upload avatar with cover
