+++ src/constants/common.ts
@@ -22,11 +22,9 @@ export const MAX_SYSTEM_ROLE_LENGTH = 9999;
 export const AVATAR_IMAGE_SIZE = 48;
 export const AVATAR_COMPRESS_SIZE = 256;
 
-export const COVER_IMAGE_WIDTH = 320;
-export const COVER_IMAGE_HEIGHT = 480;
+export const COVER_IMAGE_SIZE = 320;
 
-export const COVER_COMPRESS_WIDTH = COVER_IMAGE_WIDTH * 2;
-export const COVER_COMPRESS_HEIGHT = COVER_IMAGE_HEIGHT * 2;
+export const COVER_COMPRESS_SIZE = COVER_IMAGE_SIZE * 2;
 
 export const DEFAULT_USER_AVATAR_URL =
   'https://registry.npmmirror.com/@lobehub/assets-logo/1.2.0/files/assets/logo-3d.webp';
